+++ src/Stryker.Core/Stryker.Core/Initialisation/InputFileResolver.cs
@@ -15,7 +15,7 @@ namespace Stryker.Core.Initialisation
 {
     public interface IInputFileResolver
     {
-        ProjectInfo ResolveInput(StrykerOptions options);
+        ProjectInfo ResolveInput(StrykerOptions options, IEnumerable<IAnalyzerResult> solutionProjects);
     }
 
     /// <summary>
@@ -42,7 +42,7 @@ namespace Stryker.Core.Initialisation
         /// <summary>
         /// Finds the referencedProjects and looks for all files that should be mutated in those projects
         /// </summary>
-        public ProjectInfo ResolveInput(StrykerOptions options)
+        public ProjectInfo ResolveInput(StrykerOptions options, IEnumerable<IAnalyzerResult> solutionProjects)
         {
             var projectInfo = new ProjectInfo(_fileSystem);
             // Determine test projects
@@ -60,7 +60,7 @@ namespace Stryker.Core.Initialisation
             foreach (var testProjectFile in testProjectFiles)
             {
                 // Analyze the test project
-                testProjectAnalyzerResults.Add(_projectFileReader.AnalyzeProject(testProjectFile, options.SolutionPath, options.TargetFramework));
+                testProjectAnalyzerResults.Add(_projectFileReader.AnalyzeProject(testProjectFile, options.SolutionPath, options.TargetFramework, solutionProjects));
             }
             projectInfo.TestProjectAnalyzerResults = testProjectAnalyzerResults;
 
@@ -70,7 +70,7 @@ namespace Stryker.Core.Initialisation
             _logger.LogInformation("The project {0} will be mutated.", projectUnderTest);
 
             // Analyze project under test
-            projectInfo.ProjectUnderTestAnalyzerResult = _projectFileReader.AnalyzeProject(projectUnderTest, options.SolutionPath, options.TargetFramework);
+            projectInfo.ProjectUnderTestAnalyzerResult = _projectFileReader.AnalyzeProject(projectUnderTest, options.SolutionPath, options.TargetFramework, solutionProjects);
 
             var language = projectInfo.ProjectUnderTestAnalyzerResult.GetLanguage();
             if (language == Language.Fsharp)