+++ src/Stryker.Core/Stryker.Core/Mutants/CsharpMutantOrchestrator.cs
@@ -48,7 +48,8 @@ namespace Stryker.Core.Mutants
                 new ArrayCreationMutator(),
                 new StatementMutator(),
                 new RegexMutator(),
-                new NullCoalescingExpressionMutator()
+                new NullCoalescingExpressionMutator(),
+                new MathMutator(),
             };
             Mutants = new Collection<Mutant>();
             Logger = ApplicationLogging.LoggerFactory.CreateLogger<CsharpMutantOrchestrator>();
@@ -63,6 +64,7 @@ namespace Stryker.Core.Mutants
                     (t) => t.Modifiers.Any(x => x.IsKind(SyntaxKind.ConstKeyword))),
                 new AssignmentStatementOrchestrator(),
                 new PostfixUnaryExpressionOrchestrator(),
+                new PrefixUnaryExpressionOrchestrator(),
                 new StaticFieldDeclarationOrchestrator(),
                 new StaticConstructorOrchestrator(),
                 new PropertyDeclarationOrchestrator(),